extract operator contract document routes
